+++ app/src/main/java/com/team2/todo/screens/sub_task_details_page/SubTaskDetailsPage.kt
@@ -39,7 +39,7 @@ import java.time.LocalDateTime
 
 @OptIn(ExperimentalMaterial3Api::class, ExperimentalFoundationApi::class)
 @Composable
-fun SubTaskDetailsPage(id: Int) {
+fun SubTaskDetailsPage(id: Long) {
 
 
     val todoContext = LocalContext.current
